Moved some code around
